+++ solidity/choices/FutureAsyncChoice.sol
@@ -20,7 +20,7 @@ contract FutureAsyncChoice is AbstractAsyncChoice {
     super.activateEvent(index);
   }
 
-  function oracleCallback(uint256 correlation, bytes calldata result) external override {
+  function oracleCallback(uint16 correlation, bytes calldata result) external override {
     // Do nothing if we have already finished
     if (winner >= 0) {
       return;
